+++ src/RequestSyntaxEngine/TestFunctionality.cs
@@ -1,12 +1,13 @@
 ﻿/* PROJECT: RequestSyntaxEngine (https://github.com/aprettycoolprogram/RequestSyntaxEngine)
  *    FILE: RequestSyntaxEngine.TestFunctionality.cs
- * UPDATED: 7-1-2021-8:46 PM
+ * UPDATED: 7-8-2021-10:23 AM
  * LICENSE: Apache v2 (https://apache.org/licenses/LICENSE-2.0)
  *          Copyright 2021 A Pretty Cool Program All rights reserved
  */
 
 using System;
 using System.Reflection;
+using Utility;
 
 namespace RequestSyntaxEngine
 {
@@ -22,9 +23,7 @@ namespace RequestSyntaxEngine
                               $"MAWS Command: {RequestComponent.GetCommand(mawsRequest)}{Environment.NewLine}" +
                               $" MAWS Action: {RequestComponent.GetAction(mawsRequest)}{Environment.NewLine}" +
                               $" MAWS Option: {RequestComponent.GetOption(mawsRequest)}";
-
-            // The RequestSyntaxEngine doesn't have it's own settings file, so we hardcode the logSetting parameter.
-            Logger.Timestamped.LogEvent("trace", "TRACE", Assembly.GetExecutingAssembly().GetName().Name, logMessage);
+            LogEvent.Timestamped("trace", "TRACE", Assembly.GetExecutingAssembly().GetName().Name, logMessage);
         }
     }
 }